requirements.txt
fastapi
torch
torchvision
uvicorn[standard]
pillow
python-multipart
streamlit
requests
numpy